feat(docs): Moved Card component to components folder
